+++ R/XenaHub-class.R
@@ -121,12 +121,6 @@ XenaHub = function(hosts=xena_default_hosts(), cohorts=character(),
 ##' @return a \code{data.frame} contains all datasets information of Xena.
 ##' @author Shixiang Wang <w_shixiang@163.com>
 ##' @export
-##' @examples
-##'
-##'
-##' # XenaDataUpdate() # update newest information to local directory for use
-##' # newest_Xena = XenaDataUpdate(saveTolocal=FALSE) # just get info, not save
-##'
 XenaDataUpdate = function(saveTolocal=TRUE){
     hosts = xena_default_hosts()
     XenaList = sapply(hosts, function(x){
@@ -195,9 +189,7 @@ XenaDataUpdate = function(saveTolocal=TRUE){
 ##'
 ##' Major function of \code{UCSCXenatools}. It is used to filter
 ##' \code{XenaHub} object according to cohorts, datasets. All datasets can be found
-##' at <https://xenabrowser.net/datapages/>. Note, the change for filtering cohorts and
-##' datasets are independent.
-##'
+##' at <https://xenabrowser.net/datapages/>.
 ##'
 ##' @param x a \code{XenaHub} object
 ##' @param filterCohorts default is \code{NULL}. A character used to filter cohorts,
@@ -223,6 +215,9 @@ XenaFilter = function(x, filterCohorts=NULL, filterDatasets=NULL, ignore.case=TR
     cohorts_select = character()
     datasets_select = character()
 
+    # suppress binding notes
+    XenaHosts = XenaCohorts = XenaDatasets = NULL
+
     if (!is.null(filterCohorts)){
         cohorts_select = grep(pattern = filterCohorts, x@cohorts, ignore.case = ignore.case, value = TRUE)
     }
@@ -231,7 +226,19 @@ XenaFilter = function(x, filterCohorts=NULL, filterDatasets=NULL, ignore.case=TR
         datasets_select = grep(pattern = filterDatasets, x@datasets, ignore.case = ignore.case, value = TRUE)
     }
 
-    XenaHub(hosts = x@hosts, cohorts = cohorts_select, datasets = datasets_select)
+    if (identical(cohorts_select, character()) & identical(datasets_select, character())) {
+         warning("No valid cohorts or datasets find! Please check your input.")
+    }else{
+        if(identical(cohorts_select, character()) & !identical(datasets_select, character())){
+            UCSCXenaTools::XenaGenerate(subset = XenaHosts %in% x@hosts & XenaDatasets %in% datasets_select)
+        }else{
+            if(!identical(cohorts_select, character()) & identical(datasets_select, character())){
+                UCSCXenaTools::XenaGenerate(subset = XenaHosts %in% x@hosts & XenaCohorts %in% cohorts_select)
+            }else{
+                UCSCXenaTools::XenaGenerate(subset = XenaHosts %in% x@hosts & XenaCohorts %in% cohorts_select & XenaDatasets %in% datasets_select)
+            }
+        }
+    }
 }
 
 ##' Get hosts of XenaHub object